% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributeData.R
\name{addInstanceData}
\alias{addInstanceData}
\title{Put attribute / instance data in a NetCDF-CF File}
\usage{
addInstanceData(nc_file, names, attData = NULL)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created. It must already have an instance dimension.}

\item{names}{A character vector of names for the points.}

\item{attData}{A \code{data.frame} as included in a spatial dataFrame.}
}
\description{
Creates a NetCDF file with an instance dimension, names, and any passed
in attributes from a data frame. Use to create the start of a NetCDF-DSG file.
}
\references{
https://github.com/bekozi/netCDF-CF-simple-geometry
}

