% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPointsData.R
\name{addPoints}
\alias{addPoints}
\title{Add points to a NetCDF-DSG File.}
\usage{
addPoints(nc_file, xCoords, yCoords, alts = NULL,
  instanceDimName = "instance")
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{xCoords}{Vector of x coordinates in WGS84}

\item{yCoords}{Vector of y coordinates in WGS84}

\item{alts}{Vector of altitudes assumed to be height above mean sea level}

\item{instanceDimName}{A string to name the instance dimension.  Defaults to "instance"}
}
\description{
Adds x and y coordinates to a NetCDF-DSG dataset following the timeSeries featureType pattern.
}
\references{
https://github.com/bekozi/netCDF-CF-simple-geometry
}

