% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNCDF.R
\name{checkNCDF}
\alias{checkNCDF}
\title{Check NetCDF-DSG File}
\usage{
checkNCDF(nc)
}
\arguments{
\item{nc}{A open ncdf4 object.}
}
\description{
Introspects a netcdf file and tries to interpret it as a NetCDF-DSG file. Returns a named
\code{list} containing \code{instance_id} \code{coord_index_var} \code{coord_index_stop_var}
\code{multi_break_val} \code{hole_break_val}. If these values aren't found or aren't applicable,
they are returned \code{NULL}.
}
\references{
https://github.com/bekozi/netCDF-CF-simple-geometry
}

