% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDF.R
\name{getDF}
\alias{getDF}
\title{Get Dataframe from NetCDF-DSG File}
\usage{
getDF(nc, instance_var)
}
\arguments{
\item{nc}{A open ncdf4 object.}

\item{instance_var}{A representative 1-d instance variable.}
}
\description{
Gets instance data from a NetCDF-DSG file and returns it in a \code{data.frame}
}

