% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPoint.R
\name{multiPoint_timeSeries}
\alias{multiPoint_timeSeries}
\title{Put points in a CF timeSeries featuretype}
\usage{
multiPoint_timeSeries(nc_file, names = NULL, lats = NULL, lons = NULL,
  alts = NULL, multiPoint = NULL, add_to_existing = FALSE)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created. If NULL an object of class \code{ncdf4} is returned.}

\item{names}{A character vector of names for the points. If NULL, integers in the range 1:length(points) are used.}

\item{lats}{Vector of WGS84 latitudes}

\item{lons}{Vector of WGS84 longitudes}

\item{alts}{Vector of altitudes assumed to be height above mean sea level}

\item{multiPoint}{An object of class \code{SpatialPoints} with WGS84 lon in the x coordinate and lat in the y coordinate.
Note that MULTIPOINT Z is not supported altitude must be passed in seperately.}

\item{add_to_existing}{boolean If TRUE and the file already exists, variables will be added to the existing file.}
}
\description{
Creates point data in the NetCDF-CF timeSeries featuretype format.
}
\references{
http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html
}

