% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon_timeSeries}
\alias{polygon_timeSeries}
\title{Put polygons in a CF timeSeries featuretype}
\usage{
polygon_timeSeries(nc_file, polygonData, names = NULL,
  add_to_existing = FALSE)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{polygonData}{An object of class \code{SpatialPolygons} with WGS84 lon in the x coordinate and lat in the y coordinate.
Note that three dimensional polygons is not supported.}

\item{names}{A character vector of names for the points. If NULL, integers in the range 1:length(points) are used.}

\item{add_to_existing}{boolean If TRUE and the file already exists, variables will be added to the existing file.}
}
\description{
Creates a file with polygon instance data in the NetCDF-CF timeSeries featuretype format.
}
\references{
http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html
}

