% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToNCDFSG.R
\name{ToNCDFSG}
\alias{ToNCDFSG}
\title{Convert sp objects to NetCDF}
\usage{
ToNCDFSG(nc_file, geomData = NULL, instance_names = NULL, lats = NULL,
  lons = NULL, alts = NULL)
}
\arguments{
\item{nc_file}{A string file path to the nc file to be created.}

\item{geomData}{An object of class \code{SpatialPoints}, \code{SpatialLines} or
\code{SpatialPolygons} with WGS84 lon in the x coordinate and lat in the y coordinate.
Note that three dimensional geometries is not supported.}

\item{lats}{Vector of WGS84 latitudes}

\item{lons}{Vector of WGS84 longitudes}

\item{alts}{Vector of altitudes assumed to be height above mean sea level}

\item{names}{A character vector of names for the points. If NULL, integers are used.}
}
\description{
Creates a file with point, line or polygon instance data ready for the extended NetCDF-CF timeSeries featuretype format.
Will also add attributes if a sp dataframe object is passed in.
}
\references{
https://github.com/bekozi/netCDF-CF-simple-geometry
}

