% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrjFromNCDF.R
\name{getPrjFromNCDF}
\alias{getPrjFromNCDF}
\title{Get projection from NetCDF-CF Grid Mapping}
\usage{
getPrjFromNCDF(grid_mapping_atts)
}
\arguments{
\item{grid_mapping_atts}{A list of attributes of the grid mapping variable
as returned by ncdf or ncdf4's get attributes functions.}
}
\value{
A proj4 string for use with the sp CRS function.
}
\description{
This function takes a NetCDF-CF projection container and returns
a proj4 string.
}
\examples{
crs <- list(grid_mapping_name="latitude_longitude",
            lontidue_of_prime_meridian = 0,
            semi_major_axis = 6378137,
            inverse_flattening = 298)
prj <- getPrjFromNCDF

}
